Open sourced
Open sourced all commands. Please keep in mind that this code is old and unmaintained, and was built through the span of about 3 years, with minimal refactoring, therefore, the code here has a lot of different skill levels, built upon each other, making it almost impossible to maintain. This was rewritten latter, but the code is available in a different repository. Code here is mostly uncommented because it was for personal usage.